DRY vscode mocking in `vite.config.ts` and `vite.config.mjs` to import extensions/vscode/tests/vscode-mock.js
