+++ vite.config.ts
@@ -2,34 +2,14 @@ import yaml from '@rollup/plugin-yaml'
 import { sveltekit } from '@sveltejs/kit/vite'
 import mdsvexamples from 'mdsvexamples/vite'
 import { defineConfig, type Plugin } from 'vite'
+import { mock_vscode } from './extensions/vscode/tests/vscode-mock'
 
 export default defineConfig(({ mode }) => ({
   plugins: [
     sveltekit(),
     mdsvexamples,
     yaml(),
-    mode === `test`
-      ? {
-        // Inline virtual module to mock 'vscode' package during tests
-        name: `vscode-mock`,
-        enforce: `pre`,
-        resolveId: (id: string) => (id === `vscode` ? id : null),
-        load: (id: string) =>
-          id === `vscode`
-            ? `
-export const __noop = () => undefined
-const __proxy = new Proxy(function(){}, {
-  get: () => __proxy,
-  apply: () => undefined,
-})
-export const window = __proxy
-export const commands = __proxy
-export const workspace = __proxy
-export default { window, commands, workspace }
-`
-            : null,
-      }
-      : null,
+    mode === `test` ? mock_vscode() : null,
   ].filter((plugin): plugin is Plugin => Boolean(plugin)),
 
   test: {
